fix: fix a bug of userProfile state doesn't update when the initialState has an update.
instead of using useEffect, we can convert the page component into a seperate component, and pass that newly made component into the main component! so that we can pass in a key prop!

the key prop helps react to indentify wether the current initialState is still relevant or not. if the key prop's value changes, React will assign update the component's state even without the use of SetState!
nice.. TO HELL WITH USEEFFECT THE CAUSE OF UNNECESSARY RERENDERS >:D