refactor: add a authModalsContext Provider for all the auth modals..
rather than clutter up the navbar component, i decided to put em on a provider wrapper where i create a context there, and made it into a provider wrapper.
just wrap all of the children that would use the context, and voila. every children now has access to the authModalContext where they can access the setShow for the modals