feat: add autosave feature to create blog and edit blog, the saved values is saved on the session storage!
i implement this with a custom hook.

the hook first would set the session storage with the key that is passed from the first arg, the value for the session storage it self is passed to the second argument of this hook, we pass an object that contians the spread value of the form it self with watch method that we get from useForm hook.

the useAutoSave would resave to the session storage in an interval that is set default to 3 seconds.. it also only will save if the new value is different than the saved value on the session storage..

noice